src/main/webapp/app/entities/tipo-dni/route/tipo-dni-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { ITipoDni, TipoDni } from '../tipo-dni.model';
import { TipoDniService } from '../service/tipo-dni.service';

import { TipoDniRoutingResolveService } from './tipo-dni-routing-resolve.service';

describe('TipoDni routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: TipoDniRoutingResolveService;
  let service: TipoDniService;
  let resultTipoDni: ITipoDni | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(TipoDniRoutingResolveService);
    service = TestBed.inject(TipoDniService);
    resultTipoDni = undefined;
  });

  describe('resolve', () => {
    it('should return ITipoDni returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTipoDni = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultTipoDni).toEqual({ id: 123 });
    });

    it('should return new ITipoDni if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTipoDni = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultTipoDni).toEqual(new TipoDni());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as TipoDni })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTipoDni = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultTipoDni).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
